Started working on a SPSC Queue and more.
Dealt with memory order/ acquire and release semantics. Also, some stuff
on ABA problem, which I'm planning to make a Lockless Programming Issues
tutorial that focuses on ABA and breifly mentions data races. The major
thing is the queue, implemented as a doubly linked list, which seems like it will have issues, since it takes
quite a few operations to enqueue or dequeue something. Also, seems like
there might be an issue where if there is anything less than 2 elements in
the queue, it will have issues. Also, maybe make it a singly linked list
queue, which could probably work.