Nearly done with Lockless Queue and Locked Queue.
Worked on a Locked and Lockless version of a single producer, single
consumer queue. This is to show the ease of a locked implementation over
the complexity of a lockless implementation. Both are singly linked lists
underneath the hood. This helps demonstrate the complexity of a lockless
example fairly well, since it requires while loops that are fairly complex
and require a lot of checking.

NEED TO TIME PERFORMANCE UNDER SIMILAR
CIRCUMSTANCES.