Removed, Changed, and Added Things.
Removed the singly linked list, since the queue is implemented with one.
Removed some extranious files that I had. Moved the loading of the head
into a temporary variable statement into the mutex with the if statement
below it, since it looked like the load was not an atomic operation.
Double Check this tomorrow. Unofficially finished the Lockless Queue and
just need to finish comments in the main function to finish the tutorial.
Also look into doing a lockless heap, since singly linked list are not
often used in video games.